pyproject.toml
[tool.poetry]
name = "etherscan-converter"
version = "0.1.0"
description = ""
authors = ["djapy"]

[tool.isort]
line_length = 120
force_grid_wrap = 0
multi_line_output = 3
use_parentheses = true
include_trailing_comma = true

[tool.mypy]
exclude = [""]
follow_imports = "silent"
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
check_untyped_defs = true
no_implicit_reexport = true
disallow_untyped_defs = true
plugins = "pydantic.mypy"
ignore_missing_imports = true
show_error_codes = true
namespace_packages = true


[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.0"
opentelemetry-api = "^1.9.1"
opentelemetry-sdk = "^1.9.1"
opentelemetry-exporter-jaeger = "^1.9.1"
opentelemetry-exporter-zipkin = "^1.9.1"
opentelemetry-instrumentation-fastapi = "^0.28b1"
opentelemetry-instrumentation-aiohttp-client = "^0.28b1"
alembic = "^1.7.7"
sentry-sdk = "^1.5.7"
aiohttp = "^3.8.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.0"
pytest-asyncio = "^0.18.2"
pytest-cov = "^3.0.0"
pytest-dotenv = "^0.5.2"
flake8 = "^4.0.1"
mypy = "^0.941"
safety = "^1.10.3"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"